Unify display of a community's child communities/collections and global community list.
Community pages show a list of all direct sub-communities and all subcollections. As part of this commit, the generation of this list
is separated out into its own transformer, CommunityChildren, rather than being squashed into the CommunityViewer. The list of
sub-communities and collections is very similar to that shown on the repository front page and on the community list page, both of
which are generated by CommunityBrowser. As part of this commit, the code that generates the list is separated out into a helper class
which is then used by both CommunityBrowser and CommunityChildren.

The advantage of this approach, apart from reducing code-duplication, is that modifications such as the collapsible community list
(https://wiki.duraspace.org/display/DSPACE/Add+expanding+and+collapsing+interface+to+community-list+hierarchy+%28Manakin%29) can then
be used on community pages as well.

Squashed commit of the following:

commit 4d26438ac776c297bd6401205767a01180845fd7
Author: Andrea Schweer <schweer@waikato.ac.nz>
Date:   Wed Jul 11 17:41:35 2012 +0100

    Make depth configurable and set to 1 to match legacy behaviour

commit aee7195aff0e6e2c33f52b32645d46778a2ae511
Author: Andrea Schweer <schweer@waikato.ac.nz>
Date:   Wed Jul 11 16:37:43 2012 +0100

    Add and use new class to render community children

commit 8c9ed16553cfe3b729806cc60c4672eddd905ff4
Author: Andrea Schweer <schweer@waikato.ac.nz>
Date:   Wed Jul 11 16:36:40 2012 +0100

    Stop CommunityViewer from adding child communities/collections

commit a20f73bab801dc9147cbbde5315da60c7a649e69
Author: Andrea Schweer <schweer@waikato.ac.nz>
Date:   Wed Jul 11 16:36:03 2012 +0100

    Refactor out tree creation code for re-use elsewhere