1887 Fix citation_pdf_url behaviour
Google Scholar uses the citation_pdf_url header tag on item pages to determine
which of an item's bitstreams to link to (if any). Standard DSpace behaviour is
to include this link only if the item has exactly one bitstream in the ORIGINAL
bundle and if this item is of type application/pdf. The desired behaviour for
the LCoNZ IRRs is this:
 - if the item has only one bitstream in the ORIGINAL bundle, link to it
   (regardless of the mimetype)
 - if the item has a primary bitstream in the ORIGINAL bundle, link to it
   (regardless of how many bitstreams there are in total in the ORIGINAL bundle
   and regardless of the primary bitstream's mimetype).
Additionally, the bitstream link should only appear when the bitstream selected
according to the rules above is publicly visible.